projects/agent-friendly-git-wrapper/requirements.txt
pyyaml>=6.0
pytest>=8.0
tiktoken>=0.6.0
